~ unfix time reset at level load for Gamma
Character Select Mod "fixes" something, this "fix" breaks my mod, reeee!
The solution is obviously to break the fix. Perhaps a better solution
would be to see if ResetTime or ResetTime2 is used now, but for now
unfixing it seems fine, I am not to sure what the fix fixes apart from
redundancy.